Replace internal coroutine implementation with <https://github.com/0x5a17ed/coro>
This is a breaking change since the semantics of Generators change with
the new coroutine package. This also includes the generators returned
from the itkit/mapit package.

The old implementation would start the generator in a couroutine first and
let it run until the first time it yields a value. The new generator
implementation will start running the goroutine the moment the first
value is retrieved instead.

The automated way of closing generators was removed as well in the process
since this pattern is generally frowned upon in the Go community. The
functionality was moved into a separate package and is now available
under <https://github.com/0x5a17ed/autostop> if you are a degenerate
like me who enjoyed the automated clean up functionality.